Add a Tag and MainstreamBrowsePage model
This creates a 'tags' table in the database, and two models: Tag and
MainstreamBrowsePage. A Tag can have a parent (that is not a child), and 
we validate that the slug and title are present.

Here, single table inheritance is being used so that Tag can be of type
MainstreamBrowsePage or Topic (to be implemented later).