Add tag presenters
This adds two new presenters: TagPresenter, used to render a Tag 
instance in the required format for Panopticon, and 
MainstreamBrowsePagePresenter, which subclasses TagPresenter to override
the 'tag_type' attribute.

I've separated these concerns at this early stage as I expect there soon
to be a TopicPresenter which will also subclass the TagPresenter, but
which may include additional functionality that is currently provided
by the SectorPresenter.