Add a notifier for Panopticon
This adds a new class which encapsulates behaviour to notify Panopticon
when a tag is created, given an instance of a class that conforms to the
TagPresenter interface.

This also defines the Panopticon service, and provides an instance of
the service in development. A bearer token is always required when 
making requests to Panopticon using the API client. In development,
no assertions are made against the validity of the bearer token, as long
as there is a string present. (see the mock_bearer_token strategy in
GDS:SSO for more details)