Add a state field to Tag
This adds a new 'state' string field to the Tag model, and defines a 
state machine (using the AASM library). 

A tag is created in 'draft' by default, and can transition to the 
'published' state using the 'publish' method. We also make the 'state' 
attribute setter explicitly private, to ensure that the state can only 
be changed using the event methods.