Do not send Uncategorized groups to Publishing API
As we are adding the latest feed, we no longer need to display the
"Other" section at the bottom of a sub-topic.

This PR stops uncategorized groups, which power the "Other" section,
being sent to the Publishing API. It also adds a note to the user on the
frontend warning about this behaviour.