Add a method to return a base_path for a tag
This adds a method to the Tag class which builds the 'base path' for a 
tag instance. The method concats the slug, and the parent's slug (where
present) to build the path which the page would live on GOV.UK.

This method can be overridden by subclasses to provide their own 
prefix, as has been done with the MainstreamBrowsePage class to add the
'/browse' prefix.