Extract tag controller methods into shared module
This extracts the methods and actions in MainstreamBrowsePagesController
into a reusable module, and moves the views into a shared directory.

This will allow us to easily add a new type of tag, with similar
behaviour, in a subsequent commit.

I'm not sure if using a module here is the best idea, versus subclassing
a controller (feedback is welcome).