Restrict tag management permissions to GDS Editors
This change prevents users without the "GDS Editor" permission from
creating, updating and publishing tags. When a user without the "GDS
Editor" permission is logged in, the the "Mainstream browse pages"
navigation link is hidden.

This uses the convention we already have across other applications to
use the "GDS Editor" permission.