Add TagAssociation join table
Use an explicit join table over an implicit `has_many_and_belongs_to`
declaration so that we can control the kinds of structures we hold
references to.

Using this join table, create references `through` to the equivalent
models so that friendlier methods of `topics` and
`mainstream_browse_pages` can be used.