Add missing foreign key constraints for schema
Where possible, we should get the database to enforce the schema in
addition to our models.

This change applies a migration `AddForeignKeyConstraints`, which adds
foreign key constraints to our tables. This was done using the
`immigrant`[1] gem.

The only effective change here is in the file called
`db/migrate/20150424150725_add_foreign_key_constraints.rb`. The rest of
the diff is adding annotations for our schema structure as comments.

[1] https://github.com/jenseng/immigrant