Prevent changing the slug for all tags
We previously allowed changing the slug for draft tags.  This doesn't
work properly because it's not possible to propogate the change to
Panopticon in this way (Panopticon returns a 404 when trying to update
the tag because it can't find a tag with the new slug - obviously).

Additionally, it's not really possible to support this because the
tag_id (containing the slug) is used as the foreign key when tagging
content, and therefore updating the tag's slug would break anything
tagged with it.

We therefore can't allow editing the slugs for tags until everything has
moved over to the 'new world', and is using content_ids to link
everything.