Highlight the correct navigation item
Lists used to only belong to Topics, so whenever we were on a page
rendered by `ListsController` we highlighted the Topics entry in the
navigation.

We now support lists for mainstream browse pages too, so we need
another way to determine in which section we currently are. By adding a
helper method we can let the controllers decide which navigation item
to highlight.