Change handling of "untagged" items
The curated lists feature in this app allows users to organise links
that have been tagged elsewhere into lists.

It frequently happens that after an article is added to a list, the
article is untagged by somebody else. This means we have a `ListItem`
record in Collections Publisher that serves no purpose anymore. This is
a "normal" thing to happen.

This PR attempts to improve the layout of the section.

- Removes the separate block for the lists, as that prevents the user
from seeing the item in context.
- Shows the items in the list they originally appeared on.
- Change the button from a hard "remove" button into a "confirmation"
button. This makes it clear that we are not deleting anything
important, but just confirming a change that has been done elsewhere.

Part of improvements to the list editor. Trello:

   https://trello.com/c/VXdUyg2g/149