Send top-level pages for browse page
When collections-publisher sends a mainstream browse
page document to the content store, it should include
the full details needed to render the first two
columns of the page, so that the collections frontend
doesn't need to make a call to content-api to render
these columns of a mainstream browse page.

This information is stored in entries in the "links"
field, and consists of:

 - For `/browse`: the list of all the top level
mainstream browse pages.