Add rake task to populate Tag#published_groups
This rake task fetches the currently published `groups` (links) from
the content store and saves them in Tag#published_groups. This way we
have the currently published versions and can publish content-items
without also publishing the groups/lists.

Rake task includes a sanity check: when a non-dirty tag has a different
details.groups hash than it should have, we raise an exception because
the topic should be `dirty`. This does not happen on development.