Refactor knowledge of dependent tags
Based on pull request feedback deciding which dependent tags need to be updated has been moved to the Tag class, and should be overridden in each subclass to suit its needs.

Rather than testing the internals of the PublishingAPINotifier, expectations have been placed on the publishing_api to ensure items are being sent.