Fix N+1 query
A N+1 query was introduced recently because for each topic/browse page
on the index page, we are making a query for the subtopics with either
alphabetical or curated ordering. This caused a fair amount of extra
database traffic.

To remedy this, this commit moves the sorting of the parents to
ruby-land by using `sort_by`. This way all `children` for all tags are
eager-loaded and we just have to execute one query.