Migration to update serialised curated lists.
The Tag model stores a serialised version of the curated groups data.
This is to allow us to republish dirty tags to the content-store without
also sending the changes to the curated lists.

With the changes to the presenter to include the base_path in the groups
data, the serialised form of these needs to be updated to reflect these
changes.