Rake task to relocate detailed guides to new urls
For all published detailed guides, finds any List Items that point to
it and prefix the `base_path` with `/guidance` due to the relocation of
all detailed guides from the root namespace.

Once this has been completed, it iterates over all `Topic`s and
republishes their lists to the Publishing API via Sidekiq.

This relies on all Detailed Guides being reslugged in Rummager first,
as the `GroupsPresenter` used by `ListPresenter` relies on the content
also being tagged correctly in Rummager.