Wrap publishing in a transaction
Currently we update the database and then try to send the document to
the external services. When a downstream like whitehall fails this will
result in the item being `published?`, but the item will not have be
sent to external services.

Calling 'publish` on a published item will then result in
`AASM:​:Inva​lidTr​ansit​ion: ​Event​ 'pub​lish'​ cann​ot tr​ansit​ion
f​rom '​publi​shed'` exceptions. This has happened in production a
couple of times.