Validate that user has entered a valid GOV.UK base path
Users can enter any string into the `successor_path`, but the
content-store will only accept valid base paths. That's why we
previously reported the error "The tag couldn’t be deleted because you
didn’t enter a valid path" for all client errors. This is not ideal
because there could be other causes than the common/expected invalid
path.

To prevent users from entering an invalid GOV.UK URL we add a
validation to the form. It verifies that the input is actually a valid
base_path in the content-store. This allows us to be more vague when
the publishing-api returns a HTTPClientError.

We also clean up the controller by moving the rescuing of errors to the
form object.

Paired on this with @KushalP.