Reorder migrations to prevent migration failure
This data migration calls `Redirect#ensure_content_id_is_present` when
saving the object. That method depends on `content_id` being present.
In this ordering the `content_id` isn't added to the table yet, making
this migration fail.

The "correct" solution to these problems is to use SQL exclusively in
migration. In this case that seemed not practical because of some of
the complexity.