Remove Redirect model and create direct association between Tag and RedirectRoute
This is to fix a bug which caused a duplicate key violation when a tag was archived.
The old code attempted to create a redirect item with the same base path as the content item
that was being archived.  This code will write a redirect item with the same content id, thereby
replacing the old content_item.

Summary of changes:
- rake files to dump and restore dev database
- removal of Redirect model
- creation of direct has_many redirect_routes association on Tag
- the addition of a redirect will write details of the redirects in the content item, not as a sparate content item
- RedirectPresenter renamed as ArchivedTagPresenter to better describe its new function
- mark archived children as such in the GUI