Allow users to edit alpha taxonomy
This PR adds a section to this app where users can manage the "alpha
taxonomy". Basically it is editor to create and update content items
with format "taxon". It allows users to select a `parent` (also a
"taxon") that will be used to construct a hierarchy of taxons.

Because the publishing-api doesn't return the `links` hash for
content-items, it is currently not performant to show the hierarchy on
the index page. This might be added later.

There is also no validation yet on the position of the `parent` - a
taxon can be its own parent, or be the parent of a descendent (a loop).
We are assuming that we don't need this currently, as access to this
taxonomy will be restricted (GDS editors only).