Run all tests in schema tests
When a new PR is made on govuk-content-schemas, the Jenkins task runs
`./jenkins-schema.sh`. This test task's purpose is to run a subset of
all tests that interact with the schemas. This is done so that
govuk-content-schema tests run faster.

The problem is that for this to happen, the developer needs to tag
schema tests with `schema_test: true`. This is error prone, and we
currently have multiple schema tests that aren't triggered by jenkins.
This causes an annoying thing where a govuk-content-schemas PR is
merged that makes the tests (on master) for this repo fail.

To prevent this from happening again, we choose here to run _all_ the
tests. This causes wasted CPU cycles, but I'm of the opinion that
that's an acceptable loss compared to potential bugs not being caught
by CI. Especially now that publishing-api is working on validating the
incoming requests against the schemas.