Add browse page archiving
Currently, Collections Publisher allows users to archive or delete topics that are no longer required, but does not allow the same functionality with mainstream browse pages. The underlying functionality to do this exists in the Publishing API but is not exposed here.

This commit adds the user-facing functionality to allow mainstream browse pages to be deleted (if they have not yet been published) or archived and redirected (if they have already been published). The user experience is identical to archiving/deleting topics.