Use publishing-api to get linked items
This commit changes the way we get the pages linked to a topic or
mainstream browse page. Currently this application uses rummager, our
search index, for this. This is the case because until recently
rummager was the only application that knew about all tagging. Since
the work on the tagging infrastructure publishing-api contains all tags
(in the `links`).

By changing the data source we decrease dependency on rummager.