Send browse page parent tag to publishing-api
This application currently does not send the `parent` in the links for
browse pages, but it does for topics. This commit moves the parent (or
children) code up to the superclass so topic and browse pages act the
same.

Having the parent in the links will allow us to render the breadcrumb
from the content-store in collections.

https://trello.com/c/xvM9vb5L

After this is deployed we need to run the rake task that re-sends all
content to the publishing-api.