Fix all Rubocop linter warnings
This is in preparation for migrating the build to Jenkins 2. Jenkins 2 will require
us to add new build scripts which could optionally include Rubocop linting, so this
seems like a convenient point to add linting to this project.

All the Rubocop errors were trivial, and most could be fixed automatically.