Delete all 'beta' messages from topics
Topic pages are no longer in beta, and they will be deprecated soon when
the new taxonomy-based navigation is added. This means that the 'BETA'
tag is out-of-date, and can be deleted.

The 'beta' tag is not relevant to browse pages either, so all references
have been removed from the code.

New topics will be created with default values for properties related to
'beta' content. This currently means that they will be created in the
'live' phase, and with no 'beta' property in 'details'.