Publish route for learn to drive task list page
This commit aims to publish the route /learn-to-drive-a-car into
publishing-api.

This will be where the `task list` page for `learning to drive a car` will live within GOV.UK.

We talked with the search team and this seems to be the best way to
integrate this page with search later on.

We use the schema_name `generic` for now.

Trello:
https://trello.com/c/WOhDbn9e/208-publish-route-for-the-task-list-page