Add title and description to learn-to-drive
In https://github.com/alphagov/govuk-content-schemas/pull/720 we allow a
separate title and a new introduction field to be set on a step by step
navigation.  These will be set in the H1 on the step by step nav page and the
text immediately following.

These values are those that are currently set in the hardcoded config:
https://github.com/alphagov/collections/blob/8b9e3565e70ce1f5dedcbe501dd551198625cf93/config/step_nav/learn-to-drive-a-car.json#L4