Set up access permissions for Step by Step pages
The Modelling Services team are working on implementing the Step by Step pages into the publisher.
We want to be able to develop features and deploy them without affecting the workflow already in place.

Users with GDS editor permissions should be able to see the `Step by step pages` link next to `Topics`.