Create StepByStepPage model
The Modelling Services team are implementing the `Step by step` publishing tool into Collections Publisher.

We want to store `Step by step` content in the database, and use the data to generate the json content we want to send as part of the payload to `publishing_api`.

A second PR is required after ths one to add the `Step` model, which will have a `belongs_to` relationship with the `StepByStepPage` model.