Set up learn-to-drive step nav links
This patches links on all the learn to drive a car related content so that
a reverse link to the step by step navigation page is set up.

This allows us to move away from hard coded navigation configuration.

The commented out items in the array are pages in guides that we think of as
related, but don't have their own content ID.  They're included to avoid
confusion until the publishing tool can cope.

https://trello.com/c/nhjWvdTN/535-update-govukpublishingcomponents-to-read-sidebar-nav-data-from-links-data