Present step by step navigation page to publishing api
We want to be able to send drafts to the publishing-api when a step by step
page is saved.

We also want to discard that draft if a step by step page is deleted.

The presenter does not include edition links at the moment, but we will add
a thing to parse out links and convert them into content_ids soon.

I've also added some tests to the manage_step_by_step_page feature spec
as we were missing some.