Parse bulleted links like Govspeak
We want to accommodate content designers to continue writing their markdown
the same way was they have so far with Govspeak.

This means that both `-` and `*` can be used to indicate bulleted links.

This commit changes StepContentParser to accomodate that change.

Additionally this commit also removes LINK_CAPTURE_REGEX constant to
enable taking advantage of named capture groups in regular expressions
to use named backreferences. This is possible when the regular
expression is on the left hand side of the operator `=~`, by doing so it
assigns the captured text to local variables with the corresponding
names. [Regex Docs](https://ruby-doc.org/core-2.1.1/Regexp.html)

Trello: https://trello.com/c/2zhfNYfj