Steps use position field to reorder steps
The Modelling Services team are implementing the Step by step publishing tool into Collections Publisher.

We want to be able to reorder the `steps` of a `Step by step page` using the `position` field.

When more than one step is created and they are reordered, some javascript updates the order of the steps and populates a hidden field in the form containing an array of step `id` and `position` values. The `position` for each step is then updated in the database.