Navigation Rules Backend
StepLinksForRulesWorker is responsible to send data to run
StepLinksForRules where the links are extracted from the step contents.

Once those links are found it runs a request to Publishing API to get
the content_id of that link.

Once the content_ids are fetched from Publishing API it requests the
content to save the title, content_id and base_path in NavigationRule.

We can build up the payload for rules covering all the base paths in the current
steps.  We can then override the "include_in_links" property in the current rule
set (if any).

Then we can clear out all the rules (to ensure that we've catered for rules
that are no longer relevant), and then add in all the rules that are for
existing linked content.

This is to be later seen in the Rules page (frontend, to be done in
future commit).

Trello: https://trello.com/c/VDKP6zpq