NavigationRules & base_paths_related_to_step_nav
Also sorts the content_items alphabetically

Add two scopes into NavigationRule model
 - base_paths_part_of_step_nav
 - base_paths_related_to_step_nav

base_paths_related_to_step_nav: These are the content_items that are
related to the StepByStepPage but are not showing it on the sidebar of
their pages, it will show the related items instead.

base_paths_part_of_step_nav: These are the content_items that will have
the StepByStepPage showing on the sidebar.

The reason to have both keys is to keep a relationship between the
content_items and the StepByStepPage without having to affect the
sidebar when not wanted.

Trello: https://trello.com/c/0DUVntqs