Add publish functionality
We should only be able to publish something that has a draft
We should only be able to delete something that doesn't have a published version
I've added the skeleton of the unpublishing bits as it's coming directly next.

When you publish, you can add a change note for a minor version but it's not
mandatory.  It is mandatory for a major version.

I've started transferring to using gds-api-adapter test helpers rather than
stubbing the methods ourselves.