Check missing paths
When we link to a chapter of a guide, but not its base_path then looking up the
content_id from publishing api will not work.

For example, if we link to `/driving-lessons-learning-to-drive/taking-driving-lessons`
then the lookup for the content_id fails because the content item is registered
at `/driving-lessons-learning-to-drive`

In a lot of case this won't matter because we'll probably link to the first
page in most of the guides - which means that it will be correctly linked to.

In some cases we won't which we'll cover by checking the parent path of any
paths that we don't get a content id back for.