Add an auth_bypass_id to step-by-steps
This is the first step in allowing users to preview draft step-by-steps without a signon account.

This id will be used in future commits to
1. pass an access_limiting_token to the publishing-api payload for step-by-steps
2. create a token that users can pass into the querystring of the url to the draft preview. The token will be decoded by
[authenticating-proxy](https://github.com/alphagov/authenticating-proxy) to make sure it matches the access_limiting_token.

This code has been duplicated from [publisher](https://github.com/alphagov/publisher/blob/master/app/models/edition.rb#L431). Putting the generation of the token in a helper will make it easier to extract out to a gem later.