Amends the schema to add LinkReport table
The aim is for publishers to be able to check that links in their step-by-step guides are still alive. We're going to batch up all the links in a Step and send them to link-checker-api, collect the returned batch_id and store it against the Step, and then retrieve that batch by calling the API again when the batch is complete..
This pull request changes the schema and adds a new table to the database, called LinkReports. It also adds a relationship to the Step table. It sets up the branch for application changes.