Adds brokens links section to edit steps page.
It uses a panel and a table to display the information regarding the broken links. A table was used as there was a lot of data which is not easily comprehended via other methods. Table seems to be the most sensible way of doing things.

Ticket: https://trello.com/c/EPGwXv7z/768-add-styling-to-link-report-information-on-edit-steps-page-s