Adds method to collate links
This commit adds a method to StepContentParser to batch up all links and turn them into URLs. This makes them ready to be sent to the link-checker-api. This is the first step to implementing functionality to check whether links in Step content are still valid. They have to be parsed out and collected. I could also send a new request to the api for each link, but I think that would be inefficient.