Store publishing_app & schema_name of content_items
Store the publishing_app and schema name of the gov.uk content_item linked to in the step.

Smartanswer pages on gov.uk are made up of two content items, the smartanswer start page and the
main body of the smartanswer.
The publishing_app and schema_name values will be used in a later commit to determine whether the
content_item is a smartanswer start page, and if it is, to also get and pass the content_id of the main body
of the smartanswer to the publishing_api in the part_of_step_navs links.

`reset_column_information` is being used as a precaution to reset all of the cached columns and reload them
See the [api dock][https://apidock.com/rails/ActiveRecord/Base/reset_column_information/class] for information.