Make publishing_app and schema_name required
In a later commit a check will be added to determine whether is rule is linking
to a smartanswer start page.

This check will be used to determine whether the content_id for a smartanswer done
page should be added to the links in the payload for publishing-api.

Rather than doing truth checks on every content_item to make sure these values exist
they have been made mandatory.