Adds method to call link-checker-api with batched links
This commit adds methods and tests to the LinkReport model. The methods send off a batch of links to the link-checker-api, and then save the returned id to the database. If there are no links in the step content, then a request is never sent and the record is never created in the database.