Fixes webhook issues
This commit fixes some issues with the webhook not acting as it ought to. I have changed the route from patch to post and improved the tests so that they more closely model the data sent back from the api. I've also added `skip_before_action :verify_authenticity_token`. This was causing an error because the controller expected a CSRF token, which we don't use here because it's an API endpoint.