Add smartanswer done pages to part_of_step_nav links
Adds the smartanswer done page content_id to the part_of_step_nav links in the payload for
publishing-api if the smartanswer start page content_id is already in that list.

The smartanswer done page links are being found and added to the publishing-api
payload, right at the last minute, and are not being stored as a navigation rules
for a few reasons:

1. To remove the overhead of having to keep them up to date if the smartanswer start page is
removed from the rules.
2. So they don't need to be updated to follow suit if the step-by-step nav is hidden on the
smartanswer start page.
3. So that they don't appear as an entry on the "Choose Navigation" screen and potentially end up
with different navigation rules to the smartanswer start page.

Pages in the `related_to_step_nav` links don't display the step by step navigation, so the
smartanswer done page content_ids are not being added to this list.