Add service done pages to part_of_step_nav_links
Adds the content_id  for service done page to the part_of_step_nav of links payload for
publishing-api, if the content_id for service start page is already in that list.

There are no rules for what can be a service start page. They can be in guide, transaction
and even answer formats. As done page base_paths follow the convention of
`/done` + `start-page-slug`, I’m checking all navigations rules that aren’t a smartanswer (which does
have rules for it’s done page) to see if I can get a content_id for a done page.

I haven’t added a check to see if the done page exists, because `lookup_content_ids`
In publishing-api returns an empty list if the base_paths don’t exist.

As with smartanswer done pages, added in [PR 458](https://github.com/alphagov/collections-publisher/pull/458),
The service done page links are being found and added to the publishing-api payload, right at the last minute,
and are not being stored as a navigation rules for a few reasons:

1. To remove the overhead of having to keep them up to date if the service start page is
removed from the rules.
2. So they don't need to be updated to follow suit if the step-by-step nav is hidden on the
service start page.
3. So that they don't appear as an entry on the "Choose Navigation" screen and potentially end up
with different navigation rules to the service start page.

Pages in the `related_to_step_nav` links don't display the step by step navigation, so the
service done page content_ids are not being added to this list.

### Changes to existing tests
None of the tests except those related to done pages (smartanswer and service) care about the return
value of `lookup_content_ids`. However, having it always return something, meant that there was always
a content_id being found for the “done” page, even when one shouldn’t be found. This method now always
returns nothing unless needed.