Creates method to set `@step_by_step_page` from page
I noticed that there was a lot of repeated code. This commit moves that repeated code to a separate method. It seems that when calling methods from the page, the `params` object no longer has an `id` value. The id of the step-by-step-page is then contained in a value in `params` called `step_by_step_page_id`.