Get date most recent link check requested for step
Return the date that links where last checked for a step.

This date will be used to work out and display the last date the links were checked for
a step by step.

A note about the tests”
`links_last_checked_date` returns the date as`Tue, 07 Aug 2018 10:30:38 UTC +00:00` and
`Time.new(2018, 8, 7, 10, 30, 38)` returns the date as `2018-08-07 10:30:38 +0000` so I’m passing
both the expected and actual dates through the `utc`  method, which returns `2018-08-07 10:30:38 UTC`
to be able to compare them.