Adds edition_number column to internal_change_notes
This commit creates a new column on internal_change_notes with a default of nil. The column will be filled with the associated step by step guide's edition number, which we'll pull out of publishing-api when the step by step guide is published. This allows change notes to be associated with a particular edition. We're doing this because we've found that step by step guides may have many editions, and each edition may have many internal change notes. We think this is the best way of implementing this data hierarchy.