Ability to record and view change notes
As a content designer creating step by steps
I want to be able to see previous versions of the step by step
So that I can know who made changes to it and find out why, and so that I can revert to those versions if there's an issue with the current version

Mainstream Publisher has version history for each content item, so that you can see the changes that were made to it each time it was published. This would be very useful to have in the step by step publishing tool.

Explanation of why we're doing this work

https://trello.com/c/xurzZOox

Is there a way of measuring the effect of this work?

Name the people in the team (or others teams) that know about this work

Is there an impact for users that we need to communicate? Is there an impact for any other teams that we need to communicate?

Link to related zendesk ticket if this is a support story.

Collections Publisher

Ticket: https://trello.com/c/iDT6zgEA/824-build-interface-for-change-notes-m