Repopulate title of step nav from publishing-api
Adds the ability to reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

This functionality is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to replace the latest version
of the step nav with the last published version from publishing-api.

StepByStepPageReverter accepts any payload, as long as it’s in the format returned by publishing-api.
This means that later on the step by step publisher can be extended to allow publishers to revert back
to a specific version of the step by step.

It’s possible for the title to change between versions of the step by step, so it is being repopulated here.