Repopulate the slug of the step by step
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

It’s not possible for publishers to edit the at the moment, but this is likely to change, so the slug is being
repopulated here.

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to replace the latest version
of the step nav with the last published version from publishing-api.