Don’t repopulate the content-id for a step by step
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

The content id of the step by step should never change, so we don’t need to re-save it.

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to replace the latest version
of the step nav with the last published version from publishing-api.