Populate the position of each step
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

The number steps can be different between versions of the step by step, so they are being recreated from scratch.

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to replace the latest version
of the step nav with the last published version from publishing-api.