Saves the contents of a step that is just text
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

Repopulates the contents of a step that just contains lines of text
Adding two new lines between paragraphs, so that in the edit screen, paragraphs will be displayed with a line
between them

e.g.

```
A paragraph

Another paragraph
```

rather than

```
A paragraph
Another paragraph
```

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to