Saves bulleted lists of links with context
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

Repopulates the contents of a step that contains a bulleted list of links with context

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to