Save Navigation Rules for all internal links
Reverse engineer a step by step from publishing api and save it
as a StepByStepPage in collections-publisher.

Navigation Rules are normally created by scraping the step content for any internal links. That functionality has
been reused here.

As we are scraping the steps for links and  not checking the content items listed in
`pages_part_of_step_nav` and `pages_related_to_step_nav` to create the navigation rules, we don't have to
worry about ignoring any done pages that might appear in these lists.

Reverse engineering a step by step is the first step in allowing publishers to discard drafts of previously published
step by steps.
Rather than creating Editions in collections-publisher, we are simply going to replace the latest version
of the step nav with the last published version from publishing-api.